Refactor class structures
