+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/KeyAttestationJwt.kt
@@ -1,4 +1,159 @@
 package at.asitplus.signum.indispensable.josef
 
-class KeyAttestationJwt {
+import at.asitplus.catching
+import at.asitplus.signum.indispensable.io.InstantLongSerializer
+import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.json.JsonObject
+import kotlin.time.Instant
+
+/**
+ * Content of a Key Attestation in JWT format, according to
+ * [OpenID for Verifiable Credential Issuance](https://openid.net/specs/openid-4-verifiable-credential-issuance-1_0.html#keyattestation-jwt)
+ */
+@Deprecated("Replaced", replaceWith = ReplaceWith("KeyAttestationJwtPayload"))
+@Serializable
+data class KeyAttestationJwt(
+    @SerialName("iss")
+    val issuer: String? = null,
+
+    @SerialName("sub")
+    val subject: String? = null,
+
+    @SerialName("aud")
+    val audience: String? = null,
+
+    /**
+     * Optional. String that represents a nonce provided by the Issuer to prove that a key attestation was freshly
+     * generated.
+     */
+    @SerialName("nonce")
+    val nonce: String? = null,
+
+    @SerialName("nbf")
+    @Serializable(with = InstantLongSerializer::class)
+    val notBefore: Instant? = null,
+
+    /**
+     * Integer for the time at which the key attestation was issued using the syntax defined in RFC7519.
+     */
+    @SerialName("iat")
+    @Serializable(with = InstantLongSerializer::class)
+    val issuedAt: Instant,
+
+    /**
+     * Integer for the time at which the key attestation and the key(s) it is attesting expire, using the syntax
+     * defined in RFC7519. MUST be present if the attestation is used with the JWT proof type.
+     */
+    @SerialName("exp")
+    @Serializable(with = InstantLongSerializer::class)
+    val expiration: Instant? = null,
+
+    /**
+     * Data class containing information for instance/unit attestation
+     * which are not part of the OID4VCI specification.
+     * See https://github.com/eu-digital-identity-wallet/eudi-doc-standards-and-technical-specifications/blob/main/docs/technical-specifications/ts3-wallet-unit-attestation.md
+     */
+    @Deprecated("TS3 WUA 1.5 removed eudi_wallet_info from Key Attestations.")
+    @SerialName("eudi_wallet_info")
+    val eudiWalletInfo: EudiWalletInfo? = null,
+
+    /**
+     * Array of attested keys from the same key storage component using the syntax of JWK as defined in RFC7517.
+     */
+    @SerialName("attested_keys")
+    val attestedKeys: Collection<JsonWebKey>,
+
+    /**
+     * Optional. Array of case sensitive strings that assert the attack potential resistance of the key storage
+     * component and its keys attested in the attested_keys parameter. This specification defines initial values in
+     * Appendix D.2.
+     */
+    @SerialName("key_storage")
+    val keyStorage: Collection<String>? = null,
+
+    /**
+     * Optional. Array of case sensitive strings that assert the attack potential resistance of the user authentication
+     * methods allowed to access the private keys from the [attestedKeys] parameter.
+     * This specification defines initial values in Appendix D.2.
+     */
+    @SerialName("user_authentication")
+    val userAuthentication: Collection<String>? = null,
+
+    /**
+     * Optional. A String that contains a URL that links to the certification of the key storage component.
+     */
+    @SerialName("certification")
+    val certification: String? = null,
+
+    /**
+     * EUDI TS3 WUA 1.5: status list reference for the attested key storage and the time until which the Wallet
+     * Provider commits to maintaining the referenced status.
+     */
+    @SerialName("key_storage_status")
+    val keyStorageStatus: KeyStorageStatus? = null,
+
+    /**
+     * Optional. JSON Object representing the supported revocation check mechanisms, such as the one defined in
+     * ietf-oauth-status-list.
+     */
+    @Deprecated("TS3 WUA 1.5 replaced top-level status with key_storage_status.")
+    @SerialName("status")
+    val status: JsonObject? = null,
+) {
+
+    @Deprecated("To be removed in next release")
+    fun serialize() = joseCompliantSerializer.encodeToString(this)
+
+    @Suppress("DEPRECATION")
+    override fun equals(other: Any?): Boolean {
+        if (this === other) return true
+        if (other == null || this::class != other::class) return false
+
+        other as KeyAttestationJwt
+
+        if (issuer != other.issuer) return false
+        if (subject != other.subject) return false
+        if (audience != other.audience) return false
+        if (nonce != other.nonce) return false
+        if (notBefore != other.notBefore) return false
+        if (issuedAt != other.issuedAt) return false
+        if (expiration != other.expiration) return false
+        if (eudiWalletInfo != other.eudiWalletInfo) return false
+        if (attestedKeys != other.attestedKeys) return false
+        if (keyStorage != other.keyStorage) return false
+        if (userAuthentication != other.userAuthentication) return false
+        if (certification != other.certification) return false
+        if (keyStorageStatus != other.keyStorageStatus) return false
+        if (status != other.status) return false
+
+        return true
+    }
+
+    @Suppress("DEPRECATION")
+    override fun hashCode(): Int {
+        var result = issuer?.hashCode() ?: 0
+        result = 31 * result + (subject?.hashCode() ?: 0)
+        result = 31 * result + (audience?.hashCode() ?: 0)
+        result = 31 * result + (nonce?.hashCode() ?: 0)
+        result = 31 * result + (notBefore?.hashCode() ?: 0)
+        result = 31 * result + issuedAt.hashCode()
+        result = 31 * result + (expiration?.hashCode() ?: 0)
+        result = 31 * result + eudiWalletInfo.hashCode()
+        result = 31 * result + attestedKeys.hashCode()
+        result = 31 * result + (keyStorage?.hashCode() ?: 0)
+        result = 31 * result + (userAuthentication?.hashCode() ?: 0)
+        result = 31 * result + (certification?.hashCode() ?: 0)
+        result = 31 * result + (keyStorageStatus?.hashCode() ?: 0)
+        result = 31 * result + (status?.hashCode() ?: 0)
+        return result
+    }
+
+    companion object {
+        @Deprecated("To be removed in next release")
+        fun deserialize(it: String) = catching {
+            joseCompliantSerializer.decodeFromString<KeyAttestationJwt>(it)
+        }
+    }
 }
