+++ indispensable-josef/src/commonTest/kotlin/at/asitplus/signum/indispensable/josef/WalletAttestationJwtTest.kt
@@ -3,13 +3,14 @@ package at.asitplus.signum.indispensable.josef
 import at.asitplus.testballoon.matrix.*
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.nulls.shouldBeNull
+import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import kotlinx.serialization.json.Json
 
 val WalletAttestationJwtTest by matrixSuite {
 
     // https://github.com/eu-digital-identity-wallet/eudi-doc-standards-and-technical-specifications/blob/main/docs/technical-specifications/ts3-wallet-unit-attestation.md
-    "Wallet Instance Attestation"  {
+    "Wallet Instance Attestation" {
         val input = """
             {
               "sub": "https://client.example.com",
@@ -40,16 +41,16 @@ val WalletAttestationJwtTest by matrixSuite {
             }
         """.trimIndent()
 
-        val parsed: JwtClaims = Json.decodeFromString(input)
+        val parsed: WalletAttestationJwtPayload = joseCompliantSerializer.decodeFromString(input)
 
-        parsed.issuer.shouldBeNull()
-        parsed.walletName shouldBe "Wallet Solution X by Wonderland State Department"
-        parsed.walletVersion shouldBe "1.2.3"
-        parsed.walletLink shouldBe "https://example.com/wallet/detail_info.html"
-        parsed.walletSolutionCertificationInformation shouldBe "https://example.com/wallet/certification.html"
-        parsed.clientStatus.shouldNotBeNull()
-        parsed.confirmationClaim.shouldNotBeNull()
+        parsed.jwtClaims.issuer.shouldBeNull()
+        parsed.walletAttestationClaims.walletName shouldBe "Wallet Solution X by Wonderland State Department"
+        parsed.walletAttestationClaims.walletVersion shouldBe "1.2.3"
+        parsed.walletAttestationClaims.walletLink shouldBe "https://example.com/wallet/detail_info.html"
+        parsed.walletAttestationClaims.walletSolutionCertificationInformation shouldBe "https://example.com/wallet/certification.html"
+        parsed.walletAttestationClaims.clientStatus.shouldNotBeNull()
+        parsed.walletAttestationClaims.confirmationClaim.shouldNotBeNull()
 
-        Json.decodeFromString<JwtClaims>(Json.encodeToString(parsed)) shouldBe parsed
+        Json.decodeFromString<WalletAttestationJwtPayload>(Json.encodeToString(parsed)) shouldBe parsed
     }
 }
